+++ packages/aws-cdk-lib/core/lib/helpers-internal/index.ts
@@ -10,4 +10,4 @@ export { constructAnalyticsFromScope } from '../private/stack-metadata';
 export * from './memoize';
 export * from './reflections';
 export * from './traits';
-export { CfnPropsMixin, type CfnPropsMixinOptions } from '../mixins/private/cfn-props-mixin';
+export { CfnPropsMixin, type CfnPropsMixinOptions } from './cfn-props-mixin';
